+++ feature-libs/cart/import-export/assets/translations/en/import-export.i18n.ts
@@ -1,7 +1,7 @@
-export const exportProductList = {
+export const exportEntries = {
   exportToCsv: 'Export to CSV',
 };
 
 export const importExport = {
-  exportProductList,
+  exportEntries,
 };
