+++ feature-libs/cart/import-export/components/export-entries/export-entries.component.html
@@ -1,3 +1,3 @@
 <button class="link cx-action-link" (click)="exportToCsv()">
-  {{ 'exportProductList.exportToCsv' | cxTranslate }}
+  {{ 'exportEntries.exportToCsv' | cxTranslate }}
 </button>
