+++ feature-libs/cart/import-export/components/export-entries/index.ts
@@ -1,0 +1,2 @@
+export * from './export-entries.component';
+export * from './export-entries.module';
