--- feature-libs/cart/import-export/components/export-product-list/index.ts
@@ -1,2 +1,0 @@
-export * from './export-product-list.component';
-export * from './export-product-list.module';
