+++ feature-libs/cart/import-export/import-export.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { ImportExportCoreModule } from './core/import-export-core.module';
-import { ExportProductListModule } from './components/export-product-list/export-product-list.module';
+import { ExportEntriesModule } from './components/export-entries/export-entries.module';
 
 @NgModule({
-  imports: [ImportExportCoreModule.forRoot(), ExportProductListModule],
+  imports: [ImportExportCoreModule.forRoot(), ExportEntriesModule],
 })
 export class ImportExportModule {}
