+++ feature-libs/cart/import-export/styles/_export-entries.scss
@@ -1,4 +1,4 @@
-cx-export-product-list {
+cx-export-entries {
   display: flex;
   padding: 0 0 3rem;
 }
