+++ feature-libs/cart/import-export/styles/_index.scss
@@ -1,1 +1,1 @@
-@import './export-product-list';
+@import './export-entries';
